DEV: Skip loading login/signup controller when redirecting (#35043)
Loading the controller and its template causes a request to `hp.json`,
which can then trigger a race condition with `/session/csrf.json` if an
external auth is in-progress. This commit updates the redirect behavior
so that the the `beforeModel` hook never resolves, and therefore the
transition remains in a pending state until the redirect updates
`window.location`.

This also improves the UX, since there is no need for a flash of a
loading spinner. Instead, the normal route transition animation will
keep running until the redirect is complete.