Check for pgvector before CREATE EXTENSION
On certain hosting environments (azure postgres flex notably) CREATE EXTENSION requires a high level of permissions, which might not be available at runtime if running as a lower priv user. In such situations it's often easier to just CREATE EXTENSION manually on the database, and then run the migration. This change checks if vector already exists before enabling it.