DEV: Pin Webpack to 5.99.9 (again)
Following the update to v5.100.0, changes to files under `app/static/*`
are no longer reflected live in development mode.

Previously pinned in 712af920ce60319ba79baf4f45660e4eec39f146, but it was undone by dependabot. This time I've removed the `^` from the version specifiers, so it shouldn't be auto-upgraded any more.

Long term we're planning to move away from webpack, so it's not worth the investment to find the cause of the bug.