FIX: post voting comments editing UX (#35108)
was showing up inconditionnaly even if you didn't have the permissions
to edit other people's comments.

This was only a UX issue as the server is doing the correct check.

The issue was the `hasPermission` method which wasn't a getter so it
wasn't checked for all the instances of the post comments but rather
once.

Also added some acceptance tests to ensure we don't regress.

Internal ref - t/163347