FIX: discourse_id_challenge response when using subfolder
Added subfolder support for #discourse-id automated registration
in 874c875e02794aa but we were missing returning the "path" in
the response from the "discourse_id_challenge".

This ensures we also return the "path" field in the challenge so it can
properly be validatated by id.discourse.com.

Internal ref - t/161934/21