FIX: notification reminder is deleted with bookmark
Prior to this fix a user could delete a chat bookmark but the
notification reminder if it had already fired would still show in
notifications.

There's one notable exception: bookmarks created with the `auto_delete_preference` set to `when_reminder_sent`, given we will show the reminder at the moment we delete the bookmark, if we instantly delete the notification at the same time, it defeats the whole point of this feature.