DEV: Favor Upload.s3_upload_bucket for mediaconvert
In the aws_media_convert_adapter we were referencing
`SiteSetting.s3_upload_bucket` but that can be empty on some instances,
so we need to use the `SiteSetting.Upload.s3_upload_bucket` version,
which should always have a value when using s3.

This follows a similar pattern we used previously for other s3 related
settings.