DEV: allows to install only headless chrome
This commit allows to tweak the install of playwright chrome browser through an ENV variable: `INSTALL_CHROME_HEADLESS`.

Using a 100% headless browser is useful when running chrome in a container with llms. Cursor new browser tool for example will attempt to find the google-chrome executable to be able to navigate applications.