DEV: add timeout to video thumbnail generation
In the current version of DiscourseHub on iOS 26, video uploads are stalling.
This is because the video thumbnail generation code relies on video events
and it looks like iOS 26 webviews are not firing those events.

This may require a fix in DiscourseHub, but in the meantime, we need to
tighten the logic in thumbnail generation here.