DEV: Bypass video thumbnail generation if events don't fire (#35182)
In the current version of DiscourseHub on iOS 26, video uploads are stalling.

This is because the video thumbnail generation code relies on video events and it looks like iOS 26 webviews are not firing those events.

This may require a fix in DiscourseHub, but in the meantime, we need to tighten the logic in thumbnail generation here.