DEV: Fix `DMultiSelect` component creating one promise per input event (#35142)
Before this change, every update to the `searchTerm` tracked property would create a new promise in the `data` function. The issue was that once the debounce delay completed, only the most recent promise would be handled, leaving the other promises in a limbo state. This is problematic because we pass these promises to TrackedAsyncData, which registers a test waiter in the test environment. If promises are never settled, the test waiters will never complete, causing tests to hang indefinitely.