FIX: Confirmation prompt breaks when using pipe (#35261)
We try to prompt for a confirmation using highline before destroying posts. This by default tries to read the answer from STDIN. However, when we use a pipe, e.g. cat post_ids.txt | rake destroy:posts, STDIN is used by the pipe. This results in an error as it's already EOF.

This solves taht by initializing highline with an IO object pointing to TTY.