DEV: Simplify automation post editors (#35264)
These automation setups use a full-fledged editor where you can toggle between markdown and rich editing, but they don't have the logic that the composer comes with, that decides whether to display the preview or not, so it still (incorrectly) displays the preview when rich editor mode is selected.

This PR restricts the editor to just rich mode, and disables the preview.