FIX: add minimal load-more-sentinel height (#35262)
Based on reports in
https://meta.discourse.org/t/users-list-only-partial/368793/17, some
users may have issues with the LoadMore sentinel not triggering the
intersection observer to load more elements due to it's `0px` height.
I've not been able to reproduce this consistently myself, but adding the
`1px` height appears to resolve the issue. This is an invisible element
so it should not cause any inconsistency visually other than possibly
causing a marginally earlier load of more elements.

Credit to `cathys` on Meta Discourse for the fix.