FIX: Possible crash during category creation
In a couple of places, we were calling `site.updateCategory` with an existing category model instance. This was then passed to `.setProperties`. `setProperties` then tries to set all sorts of weird things, including overriding functions in the model like `init()`.

This commit adds an assertion on `updateCategory` to stop that happening, and updates a couple of call sites to pass a pojo instead.

Also updates the test fixture for category creation, since it was missing a load of newer attributes. Previously this wasn't noticed because a model instance was being re-used.

Co-authored-by: Alan Guo Xiang Tan <gxtan1990@gmail.com>