DEV: Make output more useful when test fails (MiniSchedulerLongRunningJobLogger) (#35293)
The test being changed some times flake with the following error
message:

```
Failure/Error: expect(fake_logger.warnings.size).to eq(1)

  expected: 1
       got: 2

```

This failure output is not useful for us and makes it harder to debug
the flake. Therefore, I've updated the assertion so the extra message
will be printed in the failure output in the future.