DEV: Remove debounce from edit sidebar category show more button (#35296)
The optimization to debounce the action when clicking on the show more
button was making our system tests harder than it needed to be since we
had to ensure we wait for the debounce. I also don't think the
performance optimisation is worth here it since it only optimises for
the case where someone clicks on multiple show more buttons really fast.