FIX: invites to private topic when using discourse connect (#35303)
When using Discourse Connect, all the invite acceptance have to go
through the SSO flow.

This worked well, unless the invite was to a "private" topic (guarded by
a group membership) which requires the "user" to click a button to
validate the invitation.

When logged in, there was a UX issue where two buttons would be
displayed

[Continue] [Accept Invite]

And the latter would not work because it's not going through the SSO
flow.

This ensures the [Accept Invite] button isn't shown in those cases.

Also refactored a type "invitiation" -> "invitation" in the
"session_controller" as well as some reworking of conditionals to
improve readability.

Ref - meta/t/384277