FIX: Safari color palette pasting (#35348)
In Safari, hitting Cmd + C or Cmd + V triggers an onKeyPress before onPaste, causing pasting colour codes to be broken.

This fixes that by adding a check for Cmd + C and Cmd + V and exiting early, deferring to the onPaste listener.

It also opportunistically fixes an issue where the colour swatches weren't circular in Safari.