FIX: Resolve race condition where draft is not loaded for a chat channel
1. In `initializers/chat-setup`, we call the async `this.chatService.loadChannels()` function.
2. `ChatChannelRoute`'s `model` calls `this.chatChannelsManager.find(params.channelId)` which will trigger an API call if `this.chatService.loadChannels`
   has not loaded the given channel yet.
3. `ChatChannel` component is rendered and `setup` function is called.
   Within the setup function,
`this.chatChannelsManager.find(params.channelId)` is called to set the
channel's draft.

If the async call in step 1 doesn't resolve before step 3 happens, the
channel draft will be blank until the component is rerendered. To
resolve this, we can move the responsibility of loading a channel's
draft into `chatDraftsManager` into the `chatChannelsManager` service
instead so that the channel's drafts are populated anytime we store a
channel into `chatChannelsManager`.