UX: Replace dropdown with a dedicated index page for email templates (#35284)
The admin email templates page currently uses a dropdown that lists all
available templates for navigating between the templates. This is an odd
way of navigation and inconsistent with how navigation is implemented in
other admin pages. This commit replaces the dropdown with a dedicated
index page that lists all templates in a table with an edit button for each
row that navigates to the edit page for the respective template.

Internal topic: t/162739.