FIX: Reload site settings before seeding content during restore
Restoring a backup executes DB migrations after the DB is restored. Part of those migrations is a step for seeding content. That step might create uploads which depends on site settings. This change excludes the seeding from the DB migrations and runs it after site settings have been restored and reloaded.