FIX: Treat the topic status enabled param as a truthy value. (#35366)
When making requests to `/t/:id/status.json`, the behaviour of the
`enabled` parameter is a bit confusing: only the string `"true"` is
treated as a true value, everything else is false (including boolean
`true`, for example).

This change makes use of the ActiveModel Boolean type to cast
falsey/truthy values to `false` and `true`.