A11Y: add role to video thumbnail button (#35326)
We markup video thumbnail `div`s in posts with `tabindex="0"` which
makes them tabbable, but without a role it's unclear they're
interactive.

Since on click or keypress these play the video, it should have
`role="button"` as well as an an accessible label.

I also noticed that `onKeyPress` was added here but isn't wired up to
anything... so I've included that and also added the space key as a
trigger in addition to enter.