A11Y: remove aria-controls from search input, fix result count (#35332)
`aria-controls` doesn't really make sense here, the input was being
associated with the number of search results... but it's not really the
expected way to use it (more for interactive controls like accordion
menus and dropdowns). This was also causing a accessibility validation
issue because `search-result-count` isn't initially present and
`aria-controls` was pointing at nothing. Removing it fixes that.

While testing this I also noticed a race condition with the result count
being announced, the count would lag and on a new search the old count
would be read by the screenreader. Clearing the count while searching
with `{{#unless @controller.searching}}` fixes the issue.