A11Y: remove aria-controls from expandable replies in topic (#35333)
This was a reasonable attempt to connect the button to the replies it
controlled (above/below a post) but the fact that the content isn't
rendered at all until expanded means we have invalid ID references and
accessibility errors. In practice this seems to cause more issues than
it solves, and we can rely on `aria-expanded` instead (already present
and properly indicates button state).