A11Y: make ariaLabelledBy conditional for topic lists (#35334)
Previously `"topic-list-heading"` was added as a reference for all topic
lists because a hidden screenreader-only H1 was added to the topic lists
on discovery pages (latest/top/hot/etc)

On other pages that use the topic list component, like suggested topics
at the bottom of topic pages, the headings are visible on-page so
`aria-labelledby` is referencing a non-existent element.

This removes `aria-labelledby` when it's not needed.