DEV: Use `TrackedArray` for reactivity in `GroupActivity` posts
Replace standard arrays with `TrackedArray` in the `GroupActivity` posts route and `PostList` component to ensure reactivity. Update the `model` method to return a tracked array and adjust the `PostList` component to handle it appropriately.

Refactor `PostList` to use `addUniqueValuesToArray` to use native array methods to add elements to the list.