FIX: Exclude fields hidden on sign-up from invite signup form (#35405)
In #34672 we added the ability to hide user fields from the sign-up form. However, we didn't exclude these from the slightly different signup form used when you are invited by another user.

This commit adds a regression test and applies the same treatment to this special version of the form.