FIX: Ensure chat service has loaded channel before rendering channel
This is a follow up to cf4193e6e1953dfe5b3b4f386c847c88f86f1c9a.

When chat is being initialized, we initiate an async request to
`/chat/api/me/channels` via `this.chat.loadChannels` but do no await on the request to be completed.
This is fine when a user is not visiting a chat channel route directly
as nothing is being rendered.

However, not awaiting on `this.chat.loadChannels` can cause problems
like a user's thread list or drafts to not be displayed if the
`/chat/api/me/channels` request does not return before rendering
happens. To resolve this, we will now wait for the promise in
`this.chat.loadChannels` on the `ChatChannelRoute` before allowing
rendering to begin.