FIX: only load the api-docs rake task definition when required gems are available (#35432)
This task was failing to load rspec-core because `lib/tasks` are loaded
in full in prod environments and this was meant to be run only in
dev/test environments (the task it is overwriting is defined in a gem
that's only loaded in those environments).

If we add more tasks like this, we can create a new subfolder in `tasks`
to segregate such tasks in a clearer manner.