FIX: Fix a loading loop when a topic is removed (#35409)
For these steps:
1. Login as a user, and read a topic;
2. The admin/moderator removes the topic;
3. Go up/down and trigger the fetch of the former/latter page.

Then the page will stuck in an infinite loading, with the console
showing `403 forbidden` getting `/t/${this.topic.id}/posts.json`. And if
the website is protected by a waf, then the user will be banned from
visiting it.

This commit fix this by adding a callback to refresh the whole page when
ajax returned 403, and then the user will see a 404 page.

<img width="881" height="612" alt="image"
src="https://github.com/user-attachments/assets/14e5db5b-cf6b-4b68-948c-44dd1de81450"
/>