FIX: eager "load more" in users directory (#35439)
When there are at least 50 users, visiting the "/u" users directory
would trigger the "load more" immediately on first load or when changing
the sorting column / direction.

This was due to a race condition between when the intersectionobserver
was running and how the sentinel was being displayed in the dom while
the rows of the users directory table was still being rendered.

The fix was to add the "isLoading" parameter that ensures the observer
is no-op-ing while the data is still being rendered.

This is different from the "enabled" parameter which prevents the
observer from... observing, because there are no more items to be loaded
for example.

Ref - https://meta.discourse.org/t/-/385110/7