UX: show only posts needing translation for default locale in AI translation chart (#35397)
The chart for AI translations included all posts for the default locale,
which would make the rest of the chart fairly disproportionate

<img width="1766" height="1228" alt="image"
src="https://github.com/user-attachments/assets/63298222-ca41-4874-8103-d5c0dae05795"
/>

In this PR I'm only showing the posts in the default locale that need
work done, which will be a much smaller number, resulting in a chart
that better represents actual translation work needed. Worst case is
that there's a very small number of posts to be translated to the
default locale, but overall this should better represent the work to be
done.

(placeholder data shown here) 
<img width="2210" height="730" alt="image"
src="https://github.com/user-attachments/assets/66e9ccb7-19f4-4aa7-addb-1b3f8f6a6dfc"
/>

I've also added a note to explain that we're not showing all the posts
in the default locale, only the ones needing translation.