FIX: remove category definition topics in hot (#35449)
In order to be consistent with other topics list (/latest, /top, ...) we
want to ensure that /hot also never shows category definition topics.

The issue was coming from that in order to help speed up /hot, we cache
the "hot score" but forgot to filter out category definition topics from
that list.

Internal ref - t/132815