FIX: phantom notification when being added to an event in a message... (#35459)
... you don't have access to.

Before creating the notification for an event, we weren't checking
whether the invitee had access to the topic, thus generating phantom
notification (aka. a number in the hamburger menu, but no notification
item in any lists).

Also made sure that creating a private event without any invitees, does
not generate invites for ... everyone 😅