DEV: Disable babel compact for theme compiler (#35469)
Compact docs:

> All optional newlines and whitespace will be omitted when generating
code in compact mode.

By default it's in auto mode, and emits a warning when it gets
auto-disabled for large files.

We use terser to strip unnecessary whitespace in production, so there's
no real benefit to us using `compact`. Easier to just disable it and
stop the warnings.