FIX: Job exception: Holidays::InvalidRegion
When we updated the list of available/supported holidays regions in
eabbac1 and in 1983a44 we left some records with invalid regions.

This adds a migration to correct old records, as well as adding an error
handling to log an error when a region is invalid, rather than throwing
an exception and breaking the background job.

Ref - https://meta.discourse.org/t/-/384873