FIX: Job exception: Holidays::InvalidRegion (#35473)
When we updated the list of available/supported holidays regions in
eabbac18cf786541268941191bffbd3070b8ee46 and in
1983a448126b36548c4cb2f60b4eeae554832cc5 we left some records with
invalid regions.

This adds a migration to correct old records, as well as adding an error
handling to log an error when a region is invalid, rather than throwing
an exception and breaking the background job.

Ref - https://meta.discourse.org/t/-/384873