FIX: prevents body scroll in input focus (#35497)
We had already a workaround for this, but this is supposed to be an even
better one. I never managed to repro the bug with this workaround.

What we are trying to prevent:
<img width="922" height="2000" alt="IMG_8116 2 (1)"
src="https://github.com/user-attachments/assets/28fb3150-1cb6-43b7-9ca6-f0e7e18dcc80"
/>

Example discussion online about this bug:

https://stackoverflow.com/questions/60797340/ios-safari-prevent-or-control-scroll-on-input-focus

This commit also extracts this logic as modifier:
`{{preventScrollOnFocus}}`