FIX: adds timeout to prevent-scroll-on-focus modifier (#35503)
The `{{prevent-scroll-on-focus}}` modifier is a workaround for a bug in
iOS where safari won't follow `preventScroll: true` and will actually
scroll. I thought that not having the timeout would be good enough, but
we actually need this delay to ensure we are past the moment where
safari will start respecting `preventScroll: true`.