FIX: Always save post drafts no matter content length
Followup 0853208d67090d4ce2a9ebec07839bfa564c16e5

Currently, whether a draft saves or not can be transparent
to members, since we don't save a draft if the title/content
of the post doesn't meet a certain length requirement. This
happens silently. We don't consider e.g. quotes in this length
calculation, so this can lead to people losing drafts, and
generally is a confusing overall experience.

In this commit, we no longer care about the length of the topic
title or post body when deciding whether to save a draft, we
will always just save.

If members end up with too many drafts they can always bulk
delete them from their profile