FIX: Skip trust level changes when group grant_trust_level is 0 (#35572)
When setting a group's Effects for trust levels to "none", the value
stored in the db for `grant_trust_level` is `0`. But since in Ruby, 0 is
not `nil`, the checks here weren't skipping the trust level
recalculation. This is specifically a problem when deleting a group with
lots of members.