DEV: AI credits should reset at start of month
## :mag: Overview

Regardless of when the `LlmCreditAllocation` was instantiated in the database, we want AI credits per LLM to reset every month. i.e. if the credits were added on the 15th of the month, it will again reset on the 1st of the month instead of the 15th of the next month